Move build_tag_dict to aws init
